easytrans/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
pycparser==2.20
PyMySQL==1.0.2
pytz==2020.4
redis==3.5.3
requests==2.25.1
ronglian-sms-sdk==1.0.0
sqlparse==0.4.1
urllib3==1.26.4
wincertstore==0.2
